Move tracing.py to internal module
